requirements.txt
pillow
transformers
torch
torchvision
torchaudio
datasets
tensorboard
protobuf<3.21.0
